refactoring: move RustType to separate module
