+++ macroslib/src/ast/collections.rs
@@ -1,65 +1,11 @@
 use proc_macro2::Ident;
 use smallvec::SmallVec;
-use smol_str::SmolStr;
-
-#[derive(Default, Debug, Clone)]
-pub(crate) struct ImplementsSet {
-    inner: SmallVec<[SmolStr; 5]>,
-}
-
-impl ImplementsSet {
-    pub(crate) fn insert(&mut self, x: SmolStr) {
-        if !self.inner.iter().any(|it| x == *it) {
-            self.inner.push(x);
-        }
-    }
-    pub(crate) fn insert_set(&mut self, o: &ImplementsSet) {
-        for it in &o.inner {
-            self.insert(it.clone());
-        }
-    }
-    pub(crate) fn contains_subset(&self, subset: &TraitNamesSet) -> bool {
-        for path in &subset.inner {
-            if !self
-                .inner
-                .iter()
-                .any(|id: &SmolStr| path.is_ident(id.as_str()))
-            {
-                return false;
-            }
-        }
-        true
-    }
-    pub(crate) fn contains(&self, trait_name: &str) -> bool {
-        self.inner.iter().any(|it| *it == trait_name)
-    }
-}
 
 struct IdentRef<'a>(&'a Ident);
 
 impl<'a> PartialEq<IdentRef<'_>> for Ident {
     fn eq(&self, o: &IdentRef<'_>) -> bool {
         self == o.0
-    }
-}
-
-#[derive(Debug, Default, PartialEq)]
-pub(crate) struct TraitNamesSet<'a> {
-    inner: SmallVec<[&'a syn::Path; 10]>,
-}
-
-impl<'a> TraitNamesSet<'a> {
-    pub(crate) fn insert<'b>(&mut self, path: &'b syn::Path)
-    where
-        'b: 'a,
-    {
-        if !self.inner.iter().any(|it| **it == *path) {
-            self.inner.push(path);
-        }
-    }
-    #[inline]
-    pub(crate) fn is_empty(&self) -> bool {
-        self.inner.is_empty()
     }
 }
